+++ src/Common/src/Interop/Unix/System.Security.Cryptography.Native/StreamSizes.cs
