+++ src/Common/src/System/Net/Security/Unix/SafeFreeCredentials.cs
@@ -1,0 +1,74 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+using  Microsoft.Win32.SafeHandles;
+
+namespace System.Net.Security
+{
+    //
+    // Implementation of handles dependable on FreeCredentialsHandle
+    //
+#if DEBUG
+    internal abstract class SafeFreeCredentials : DebugSafeHandle
+    {
+#else
+    internal abstract class SafeFreeCredentials : SafeHandle
+    {
+#endif
+        protected SafeFreeCredentials(IntPtr handle, bool ownsHandle) : base(handle, ownsHandle)
+        {
+        }
+    }
+
+    //
+    // This is a class holding a Credential handle reference, used for static handles cache
+    //
+#if DEBUG
+    internal sealed class SafeCredentialReference : DebugCriticalHandleMinusOneIsInvalid
+    {
+#else
+    internal sealed class SafeCredentialReference : CriticalHandleMinusOneIsInvalid
+    {
+#endif
+
+        //
+        // Static cache will return the target handle if found the reference in the table.
+        //
+        internal SafeFreeCredentials Target;
+
+        internal static SafeCredentialReference CreateReference(SafeFreeCredentials target)
+        {
+            SafeCredentialReference result = new SafeCredentialReference(target);
+            if (result.IsInvalid)
+            {
+                return null;
+            }
+
+            return result;
+        }
+        private SafeCredentialReference(SafeFreeCredentials target) : base()
+        {
+            // Bumps up the refcount on Target to signify that target handle is statically cached so
+            // its dispose should be postponed
+            bool ignore = false;
+            target.DangerousAddRef(ref ignore);
+            Target = target;
+            SetHandle(new IntPtr(0));   // make this handle valid
+        }
+
+        protected override bool ReleaseHandle()
+        {
+            SafeFreeCredentials target = Target;
+            if (target != null)
+            {
+                target.DangerousRelease();
+            }
+
+            Target = null;
+            return true;
+        }
+    }
+}
